chore: try esm
